refac: separate model classes
