+++ models/entity.ts
@@ -1,4 +1,6 @@
 import { Module } from 'wollok-ts/dist/model'
+import { Method } from './method'
+import { Attribute } from './parameter'
 
 export type Kind = Module['kind']
 
@@ -16,29 +18,5 @@ export class Entity {
 
 	addMethod(method: Method) {
 		this.methods.push(method)
-	}
-}
-
-export class Method {
-	constructor(public name: string, public parameters: string[]) {}
-
-	get description() {
-		return `${this.name}(${this.parameters.join(', ')})`
-	}
-
-	get key(){
-		return this.name
-	}
-}
-
-export class Attribute {
-	constructor(public name: string, public constant: boolean, public property: boolean) {}
-
-	get description() {
-		return this.name
-	}
-
-	get key(){
-		return this.name
 	}
 }
